+++ packages/builder/builder-rspack-provider/src/core/createCompiler.ts
@@ -1,11 +1,11 @@
 import {
   debug,
   logger,
+  prettyTime,
   formatStats,
   TARGET_ID_MAP,
 } from '@modern-js/builder-shared';
 import type { Context, RspackConfig } from '../types';
-import prettyTime from '@modern-js/builder-shared/pretty-time';
 
 export async function createCompiler({
   context,
@@ -35,7 +35,7 @@ export async function createCompiler({
     if (!stats.hasErrors()) {
       obj.children?.forEach((c, index) => {
         if (c.time) {
-          const time = prettyTime([0, c.time * 10 ** 6], 0);
+          const time = prettyTime([0, c.time * 10 ** 6]);
           const target = Array.isArray(context.target)
             ? context.target[index]
             : context.target;