+++ packages/builder/builder-webpack-provider/src/webpackPlugins/ProgressPlugin/ProgressPlugin.ts
@@ -1,6 +1,6 @@
 import webpack from 'webpack';
 import { logger } from '@modern-js/utils/logger';
-import prettyTime from '@modern-js/builder-shared/pretty-time';
+import { prettyTime } from '@modern-js/builder-shared';
 import { bus, createFriendlyPercentage } from './helpers';
 import { createNonTTYLogger } from './helpers/nonTty';
 import type { Props } from './helpers/type';
@@ -85,7 +85,7 @@ export class ProgressPlugin extends webpack.ProgressPlugin {
     compiler.hooks.done.tap(this.name, stat => {
       if (startTime) {
         this.hasCompileErrors = stat.hasErrors();
-        this.compileTime = prettyTime(process.hrtime(startTime), 0);
+        this.compileTime = prettyTime(process.hrtime(startTime));
         startTime = null;
 
         if (!this.hasCompileErrors) {