+++ packages/builder/builder-cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/builder-cli
 
+## 2.35.1
+
+### Patch Changes
+
+- Updated dependencies [ea3fe18]
+- Updated dependencies [9dd3151]
+- Updated dependencies [bc1413d]
+- Updated dependencies [0988b12]
+- Updated dependencies [4980480]
+- Updated dependencies [dc93952]
+- Updated dependencies [6b03be8]
+- Updated dependencies [578f94f]
+- Updated dependencies [6a1d46e]
+- Updated dependencies [fe8caeb]
+  - @modern-js/builder-webpack-provider@2.35.1
+  - @modern-js/builder-rspack-provider@2.35.1
+  - @modern-js/builder@2.35.1
+  - @modern-js/utils@2.35.1
+
 ## 2.35.0
 
 ### Patch Changes